+++ geotrek/sensitivity/models.py
@@ -96,6 +96,9 @@ class SensitiveArea(MapEntityMixin, StructureRelated, TimeStampedModelMixin, NoD
         db_table = 's_t_zone_sensible'
         verbose_name = _(u"Sensitive area")
         verbose_name_plural = _(u"Sensitive areas")
+        permissions = (
+            ("import_sensitivearea", "Can import Sensitive area"),
+        )
 
     def __unicode__(self):
         return self.species.name
@@ -193,5 +196,5 @@ class SensitiveArea(MapEntityMixin, StructureRelated, TimeStampedModelMixin, NoD
     pretty_practices_verbose_name = _("Practices")
 
 
-Topology.add_property('sensitive_areas', lambda self: intersecting(SensitiveArea, self), _(u"Sensitive areas"))
-Topology.add_property('published_sensitive_areas', lambda self: intersecting(SensitiveArea, self).filter(published=True), _(u"Published sensitive areas"))
+Topology.add_property('sensitive_areas', lambda self: intersecting(SensitiveArea, self, settings.SENSITIVE_AREA_INTERSECTION_MARGIN), _(u"Sensitive areas"))
+Topology.add_property('published_sensitive_areas', lambda self: intersecting(SensitiveArea, self, settings.SENSITIVE_AREA_INTERSECTION_MARGIN).filter(published=True), _(u"Published sensitive areas"))